+++ mocks/include/mocks/EmptyApplication.hpp
@@ -104,6 +104,12 @@ public:
         return nullptr;
     }
 
+    PubSub *getTwitchPubSub() override
+    {
+        assert(false && "getTwitchPubSub was called without being initialized");
+        return nullptr;
+    }
+
     Logging *getChatLogger() override
     {
         assert(!"getChatLogger was called without being initialized");